+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "mmap-value-storage"
+name = "blob_store"
 version = "0.1.0"
 edition = "2021"
