+++ rocksdb-benches/Cargo.toml
@@ -8,7 +8,7 @@ rocksdb = ["dep:rocksdb"]
 
 [dependencies]
 bustle = "0.5.1"
-mmap-value-storage = { path = ".." }
+blob_store = { path = ".." }
 parking_lot = "0.12.3"
 rocksdb = { version="0.22.0", optional=true }
 serde_json = "1.0.132"
