+++ src/blob.rs
@@ -1,4 +1,4 @@
-pub trait Value {
+pub trait Blob {
     fn to_bytes(&self) -> Vec<u8>;
 
     fn from_bytes(bytes: &[u8]) -> Self;
